pyproject.toml
[project]
name = "triggers-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Robert Hernandez", email = "hernnadez20robert@gmail.com" }
]
requires-python = ">=3.14.2"
dependencies = [
    "mysql-connector-python>=9.6.0",
    "pyqt6>=6.10.2",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
